% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_rcc_dependencies}
\alias{install_rcc_dependencies}
\title{Install rcc dependencies}
\usage{
install_rcc_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{method}{character string passed on to argument of the same name in \link[reticulate]{py_install}.}

\item{conda}{character string passed on to argument of the same name in \link[reticulate]{py_install}.}
}
\value{
Nothing. Libraries are installed in the current python environment.
}
\description{
\code{install_rcc_dependencies} installs or updates numpy and scipy.
}
\examples{
\dontrun{
# setup -----

# create and set environment
reticulate::conda_create("r-reticulate")
reticulate::use_condaenv("r-reticulate")

# install rcc dependencies
install_rcc_dependencies()
}

}
