% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{Edgar Anderson's Iris Data}
\format{
\code{iris} is a data frame with 150 cases (rows) and 5 variables (columns) named \code{Sepal.Length}, \code{Sepal.Width}, \code{Petal.Length}, \code{Petal.Width}, and \code{Species}.
}
\usage{
iris
}
\description{
Copied from \link[datasets]{iris}
}
\references{
Edgar (1935). The irises of the Gaspe Peninsula, Bulletin of the American Iris Society, 59, 2–5.
}
\keyword{datasets}
