% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rcc_environment}
\alias{setup_rcc_environment}
\title{Setup rcc environment}
\usage{
setup_rcc_environment(verbose = TRUE)
}
\arguments{
\item{verbose}{logical controlling verbosity of function.}
}
\value{
Nothing. Conda environment is created and activated.
}
\description{
\code{setup_rcc_environment} creates and activates Rrcc environment.
}
\examples{
\dontrun{
setup_rcc_env()
}

}
